# src/main/resources/functional_requirements/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/ui/payment/{id}": {
      "get": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "getPaymentById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataPayment"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "updatePayment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "transition",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Payment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataPayment"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "deletePayment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/ui/payment": {
      "get": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "getAllPayments",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityWithMetadataPayment"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "createPayment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Payment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataPayment"
                }
              }
            }
          }
        }
      }
    },
    "/ui/payment/search/advanced": {
      "post": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "advancedSearch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityWithMetadataPayment"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/ui/loan/search/advanced": {
      "post": {
        "tags": [
          "loan-controller"
        ],
        "operationId": "advancedSearch_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoanSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityWithMetadataLoan"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/ui/payment/search/loan/{loanId}": {
      "get": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "getPaymentsByLoan",
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityWithMetadataPayment"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ui/payment/business/{paymentId}": {
      "get": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "getPaymentByBusinessId",
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataPayment"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Payment": {
        "type": "object",
        "properties": {
          "paymentId": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "valueDate": {
            "type": "string",
            "format": "date"
          },
          "currency": {
            "type": "string"
          },
          "sourceReference": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "loanId": {
            "type": "string"
          },
          "matchingReference": {
            "type": "string"
          },
          "virtualAccountReference": {
            "type": "string"
          },
          "allocation": {
            "$ref": "#/components/schemas/PaymentAllocation"
          },
          "paymentFileId": {
            "type": "string"
          },
          "sourceType": {
            "type": "string"
          },
          "capturedAt": {
            "type": "string",
            "format": "date-time"
          },
          "capturedBy": {
            "type": "string"
          },
          "matchedAt": {
            "type": "string",
            "format": "date-time"
          },
          "allocatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "postedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          }
        }
      },
      "PaymentAllocation": {
        "type": "object",
        "properties": {
          "interestAmount": {
            "type": "number"
          },
          "feesAmount": {
            "type": "number"
          },
          "principalAmount": {
            "type": "number"
          },
          "totalAllocated": {
            "type": "number"
          },
          "unallocatedAmount": {
            "type": "number"
          },
          "allocationNotes": {
            "type": "string"
          }
        }
      },
      "EntityMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "modelKey": {
            "$ref": "#/components/schemas/ModelSpec"
          },
          "state": {
            "type": "string"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "transitionForLatestSave": {
            "type": "string"
          }
        }
      },
      "EntityWithMetadataPayment": {
        "type": "object",
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/Payment"
          },
          "meta": {
            "$ref": "#/components/schemas/EntityMetadata"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string"
          },
          "modelKey": {
            "$ref": "#/components/schemas/ModelSpec"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "transitionForLatestSave": {
            "type": "string"
          }
        }
      },
      "ModelSpec": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Loan": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "string"
          },
          "partyId": {
            "type": "string"
          },
          "principalAmount": {
            "type": "number"
          },
          "apr": {
            "type": "number"
          },
          "termMonths": {
            "type": "integer",
            "format": "int32"
          },
          "dayCountBasis": {
            "type": "string"
          },
          "repaymentDay": {
            "type": "integer",
            "format": "int32"
          },
          "fundedDate": {
            "type": "string",
            "format": "date"
          },
          "fundedAmount": {
            "type": "number"
          },
          "outstandingPrincipal": {
            "type": "number"
          },
          "accruedInterest": {
            "type": "number"
          },
          "totalInterestReceivable": {
            "type": "number"
          },
          "maturityDate": {
            "type": "string",
            "format": "date"
          },
          "nextDueDate": {
            "type": "string",
            "format": "date"
          },
          "referenceSchedule": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoanScheduleItem"
            }
          },
          "approval": {
            "$ref": "#/components/schemas/LoanApproval"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          }
        }
      },
      "LoanApproval": {
        "type": "object",
        "properties": {
          "approvedBy": {
            "type": "string"
          },
          "approvedAt": {
            "type": "string",
            "format": "date-time"
          },
          "approverRole": {
            "type": "string"
          },
          "approvalComments": {
            "type": "string"
          },
          "checkedBy": {
            "type": "string"
          },
          "checkedAt": {
            "type": "string",
            "format": "date-time"
          },
          "checkerRole": {
            "type": "string"
          }
        }
      },
      "LoanScheduleItem": {
        "type": "object",
        "properties": {
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "principalAmount": {
            "type": "number"
          },
          "interestAmount": {
            "type": "number"
          },
          "totalAmount": {
            "type": "number"
          },
          "remainingBalance": {
            "type": "number"
          }
        }
      },
      "EntityWithMetadataLoan": {
        "type": "object",
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/Loan"
          },
          "meta": {
            "$ref": "#/components/schemas/EntityMetadata"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string"
          },
          "modelKey": {
            "$ref": "#/components/schemas/ModelSpec"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "transitionForLatestSave": {
            "type": "string"
          }
        }
      },
...
      "PaymentSearchRequest": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "string"
          },
          "minAmount": {
            "type": "number"
          },
          "maxAmount": {
            "type": "number"
          },
          "valueDateFrom": {
            "type": "string",
            "format": "date"
          },
          "valueDateTo": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "LoanSearchRequest": {
        "type": "object",
        "properties": {
          "partyId": {
            "type": "string"
          },
          "minPrincipal": {
            "type": "number"
          },
          "maxPrincipal": {
            "type": "number"
          },
          "termMonths": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}